--- ui/apps/dev-server-ui/src/app/(dashboard)/apps/(onboarding)/choose-template/templates.json
@@ -1,10 +1,0 @@
-{
-  "templates": [
-    {
-      "name": "Astro",
-      "logo": "",
-      "url": "https://github.com/inngest/inngest-js/tree/main/examples/framework-astro",
-      "sdk_language": "TypeScript"
-    }
-  ]
-}
